libhwc2.1: Move ExynosLayer::checkFps() call time
This patch moves ExynosLayer::checkFps() call time to
the ExynosLayer::setLayerBuffer().
There are below two reasons.

- Previous code had bug in calculating layer's fps.
HWC increased ExynosLayer::mFrameCount twice when
handleLowFpsLayers or dynamic recomposition is enabled
because checkFps() was called twice.
This patch calls checkFps() once when layer buffer
is set.

- mGeometryChanged flag should be set before next
validateDisplay or presentDisplay is called when
fps is chaged in order to enable skipping resource assignment
and skipping validate feature.

Change-Id: Ie83dd0e0b3580c1ba892696119575adc5bdf4a56
Signed-off-by: HyunKyung Kim <hk310.kim@samsung.com>