libhwjpeg: fix APP1 size calculation
libhwjpeg calculate the length of APP1 before writing data to resereve
the space for embedded thumbnail jpeg. However the calcluation logic
does not consider the data that can be stored in both of offset field
and separate data area.
For example, if an ASCII entry has 3-byte data, it should be written
in offset field. The size of the field and its data should be 12 bytes
but libhwjpeg treats it as 15 bytes which consists of 12-byte entry
and 3-byte data written after its IFD.
If the calculation is not correct, libhwjpeg consequently updates the
length of APP1 segment incorrectly by
CAppMarkerWriter::UpdateAPP1Size().

Change-Id: I661a208ae415aa53150c93d8f0bb1204c38f9e92
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>