libhwjpeg: change the order of quantization table
ExynosJpegEncoder::setQuality() allows users to specify two
quantization tables for Luma and Chroma planes. It requires the
quantizers in the talble shoud be specified in raster-scan order when
it is first instroduced because it is natural to human.
But CHWJpegCompressor::SetQuality() requires the quantization table to
be specified in zigzag scan order because it is natural to the people
familiar with HWJPEG and the JPEG specifications (ITU-t81). The DQT in
JPEG also specifies the quantizers in zig-zag scan order.
Now we decided to change the arrangement of quantizers to
ExynosJpegEncoder::setQuality() not to confuse the users of libhwjpeg.

Both ExynosJpegEncoder::setQuality() and CHWJpegCompressor::SetQuality()
now need the quantizers to be specified in zig-zag scan order.

Change-Id: I284f36d9dcc1f5af0dec92fdc0e1a3430e09e3a2
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>