libacryl: add new ioctl for extended commands
The old g2d_task structure passed commands of target
as an array. However, commands can be added according
to H/W changes. we add new ioctl to use extended
argument that passed commands of target as a pointer.

To support existing G2D driver, we add the version of
H/W to make acrylic class. If version of H/W is lower than
G2D 2020 version, we uses existing ioctl for compatibility.

Change-Id: I596e4baf1e52568fda72aa6e418a773a16a71886
Signed-off-by: hyesoo.yu <hyesoo.yu@samsung.com>
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>