libacryl: deliver layer opaque data
If users want to attach an opaque data associated with its layer, they
can deliver it to the layer with setLayerData(). Since it needs to a
pointer to some data, the memory that contain the data should be
managed by the caller properly until the data is cleared explicitely
with clearLayerData().

Change-Id: I123dcc3ded1932313873276ae75c3f1b08173d4c
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>