libacryl: change ycbcr conversion siting value
There is a color problem in case of color space conversion
of odd position. For YUV420 or YUV422, since each U or V sample
is used to two pixels horizontally, the odd position pixel
chroma is used as an even numbered chroma value.
Thus, we enable to interpolate two pixel to fix color change.

Change-Id: I21b60943eda2aa9cafead0198946292c0c2d8c83
Signed-off-by: Hyesoo Yu <hyesoo.yu@samsung.com>