libhwjpeg: refactor Set[Src|Dst]Image()
The LibScalerForJpeg is the thumbnail image generator for libhwjpeg and
the image and buffer types to LibScalerForJpeg does not vary.
Unnecessary flexibility increases the complexity of a module. So, we
should discard the support for userptr in SetDstImage().
In addition, LibScalerForJpeg requires unnecessary type casting to
libhwjpeg for SetSrcImage and SetDstImage for the historical reasion.
However, it is no longer needed because we can decide our API now.

Change-Id: Ie8eebc38bf4f09889cc001e673f128570fb642c1
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>