libhwjpeg: delegate driver init to each port
V4L2 API call sequence in LibScalerForJpeg is deviced by the APIs
for both of the ports including output and capture:
  s_fmt(output) & s_fmt(capture) -> reqbufs(output) & reqbufs(capture)
  -> qbuf(output) & qbuf(capture) -> dqbuf(output) & dqbuf(capture).
But LibScalerForJpeg should revert the success of an API call for
output if the following the same API call for capture is failed for
the next try.
Now we can choose the better way:
  (s_fmt(output)->reqbufs(output))&(s_fmt(capture)->reqbufs(capture))
  -> qbuf(output) & qbuf(capture)
We do not need to revert an API call for output even though its
following API call for capture.o

We also moved parameters of LibScalerForJpeg::Set[Src|Dst]Image()
about image buffers to LibScalerForJpeg::RunStream() because image
dimension and format seldom change while buffers can be changed every
frame. This change helps the better design. We do not need to hold the
buffer information confirued by LibScalerForJpeg::Set[Src|Dst]Image()
until LibScalerForJpeg::RunStream() is called.

Change-Id: I4af0750975068b22055c1c2eafc53ce76d366479
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>