libhwjpeg: delegate qbuf to each port
If v4l2_buffer instances are the members of LibScalerForJpeg, it is
hard to understand where each field of the instances is initialized.
So, it had better allocate a v4l2_buffer instance in the stack and
initialze it everytime we need to call V4L2_QBUF and V4L2_DQBUF.
This change introduces extra overhead for initializing all fields of
the instance but makes LibScalerForJpeg simpler. The overhead is very
small compared to the overhead by other tasks performed by
LibScalerForJpeg.

Change-Id: Icb13c16f0f880ecd572d3461a2b71dcb321d1fb8
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>