libhwjpeg: introduce polymorphic thumbnail scaler
LibScalerForJpeg is the only thumbnail scaler implemented with the
legacy V4L2 MSCL driver. But we may need to use another types of
thumbnail scaler that runs with different H/W or different API.
Therefore abstract calss, ThumbnailScaler is introduced and the
concrete object of the thumbnail scaler is created with
ThumbnailScaler::createInstance(). All the implementation details
and differences between products are handled in createInstance().

Change-Id: Ibef2693ca6ea9e23f8993c9e525e58ac54a22333
Signed-off-by: Cho KyongHo <pullip.cho@samsung.com>