libhwc2.1: Fix bug in checking dst align resriction
Dst position and size should be aligned with 16 if
destination image is AFBC compressed and background
layer is not used.
This patch applies this restriction.
mNeedCompressedTarget rather than dst.compressed
indicates whether dst buffer is compressed or not.
Previous code used dst.compressed to check whether
dst buffer is compressed. This patch fixed this bug.

Bug: 143335430

Change-Id: I439f2d9ab2144d886238c872696bd4ae84e8db5b
Signed-off-by: HyunKyoung <hk310.kim@samsung.com>