libhwc2.1: Correct the handle color format when reuse destination buffer
MPP tends to reuse destination buffers when the buffer size is bigger
than requested. However, it created the inconsistency between the buffer
info and its use in MPP. The patch has the first step to resolve the
inconsistency inside the buffer format. To ensure Gralloc 4.0 can work,
we need to review all metadata info we will use.

Bug: 170821166
Test: Run resize animation for YouTube mini-player.
Change-Id: I274bfa8445cd3ba2b66263d7c6e36320396e3cc2