Move the IdleHintWorker to the end.
The IdleHintWorker must be the last field in the ExynosDisplay
struct. When the ExynosDisplay is destroyed, the thread on which the
Routine() function runs must be destroyed before any of the other
fields because it may otherwise observe the ExynosDisplay object in
a partially destructed state. The destructor destroys fields in
reverse order of their declarations, so having this field be last
ensures this property.

Bug: 170327170
Change-Id: I5f293079eec9425429e6a4a6c1aa03ff3432f048