libhwc2.1: Separate plane type to support new purpose plane
In case of GS101, the number of planes is same as the number of windows.
But, new purpose planes are added for RCD (Round Corner Display)
function from GS201 and it occurs some problems. This patch
separates planes to normal planes which are used for window
composition and special purpose planes. The number of windows will
be calcurated from the number of normal planes.

Bug: 186553202
Signed-off-by: YongWook Shin <yongwook.shin@samsung.com>
Change-Id: I1a3048c86f2479da974c6be7f63ade1342751a09