libhwc2.1: fix idle hint checking before waiting signal
Before calling WaitForSignalOrExitLocked(), we need to ensure
mIdleHintIsEnabled is correct according to mIdleHintDeadlineTime
and the current time.

This will also make sure that the worker has to check the idle hint
support immediately (not waiting for the signal first) and will
always retry if there's a failure when checking the support.

Bug: 189389118
Test: 1. Simulate signalIdle() is called right before Lock(), and
         idle hint signals should be handled immediately.
      2. Simulate delaying some earliest signals, and idle hint
         support should be checked immediately without waiting for
	 a signal first, and if there's a failure, it will retry.
Change-Id: I2b75c595540bd284463e301a0e13da5e7732b6ac