libhwc2.1: send idle hint while video playback
We will send a display idle hint if there is no frame update for
100ms. For the case of video playback, we will not send the hint
since the composer keeps updating the frame. This change ignores
the update of video layers and only considers UI layers for sending
the hint in a frame, i.e. we will send the idle hint if we have
video layers but UI layers are not updated so that we can save more
power in video playback.

Bug: 199292504
Test: capture a trace to confirm the idle hint behavior in video
      playback
Change-Id: I9bb4fe69a4089fc1c7154f89f01b7b8f05208361