libhwc2.1: invalidate mTargetBuffer after ComposerClient has died
We need to invalidate the client target buffer when ComposerClient
has died because it will automatically free the client target buffer.
We also need to invalidate the skip static layer flag so that we
have to get a new client target buffer first, before we can skip the
static layers.

We can use unregisterEventCallback_2_4 as a sign of ComposerClient's
death since it will be called during the destructor of
~ComposerClientImpl.

Bug: 188745089
Test: kill SF process, and mTargetBuffer should be invalidated
Test: mTargetBuffer should not be accessed after it has been
      invalidated.
Change-Id: I4b12b6fa96dbd5a1538a91e718e96d33167520a8