libhwc2.1: don't retry checking idle hint support on start
We don't need to repeat checking the idle hint support when the
support checking is failed on the start since it will cause a
slowdown of the boot time. This could happen when power hal is
not ready yet.

But, at least we will still retry to check the support whenever
there's a frame update or there's a refresh rate update.

Test: atp:asit/perf/boottime_test
Bug: 206760351

Change-Id: Ib338a467d639405d761bf7c21b8cbae936d72601