libhwc2.1: avoid pure virtual function call in Worker thread
We should not call InitWorker() inside PowerHalHintWorker's constructor
since InitWorker() will spawn a worker thread that will call virtual
function Routine() meanwhile PowerHalHintWorker's constructor is not
finished yet.

We also need to move Exit() from Worker's destructor to its derived
class's destructor so that the thread will be closed before the worker
object becomes in partially destructed state.

This will also include the fix for needUpdateIdleHintLocked() so that
the condition for enabling idle hint is exactly the same as
updateIdleHint().

Test: Routine() should only be called after PowerHalHintWorker's ctor
	has been finished.
Test: Routine() should not be called repeatedly without any reason when
	there's still no frame update in the beginning.
Bug: 207791724
Change-Id: I44f6b3829c0bc5364d1fa9b1880e9f903ba2f017