libhwc2.1: wait for peak refresh rate before enabling LHBM
We have to switch to the peak refersh rate (e.g. 90/120Hz) before
enabling LHBM due to panel constraints. The peak refresh rate is
set by the fingerprint but it doesn't make sure we are running at
the peak. Wait for the peak according to mConfigRequestState and
mVsyncPeriod to guarantee we can enable LHBM.

Bug: 233145265
Test: use fingerprint to unlock the device, auto test by QMC team
Change-Id: I794ca743101f3463596a25bde0127ec9920a3658