libhwc2.1: correct config state and vsync period
If there is a pending config in AOD mode or change to same config
rapidly, the config state will be REQUESTED. It should reset to NONE
in those cases. Besides, vsync callback may get incorrect mVsyncPeriod
if it refers to mActiveConfig while quickly switching fps. To avoid it,
refer to mActiveModeState to update mVsyncPeriod.

Bug: 232044740
Test: Suspend/resume, AoD, quickly fps switch
Change-Id: Ifc8989256a9b6f6633abc2d4d74124874493c6c0