libhwc2.1: reassign sandwiched layer that is not opaque
To keep sandwiched layer to DEVICE composition, composer request SF to
clear client target with transparent pixels where this sandwiched layer
region by setting HWC2_LAYER_REQUEST_CLEAR_CLIENT_TARGET. However, it is
not workable for layer is not opaque, because the layer must blend with
whatever is underneath. Therefore we need to reassign sandwiched layer
to CLIENT composition if it is not opaque.

Bug: 253304752
Test: front buffer test, play video
Change-Id: I39c8e3430fee6cd4149e85f510ce2900ff46c3d0