libhwc2.1: always update log files with the latest info
Previously, the log files would not be updated anymore if the size is
already more than 1 MB. But now, the log files should always be updated
but the total size and the file count are still guaranteed to be
limited.

This will also change the log files to be per display (to avoid race
condition for multi-display).

Bug: 245511651
Test: make many errors and check HWC log files
Test: vndservice call Exynos.HWCService 11 i32 0 i32 308 i32 1
Change-Id: I3f3f0ead7bb5b513d51388a13fcb67762031d93e