libhwc2.1: update preassignedDisplay information
All MPP resources have preassignedDisplay information at initial
time. It means each MPP resource can bind with one of Primary,
Secondary and External Display.
When Secondary Display is not valid, the MPP resources which
are binded with it can be assigned to Primary or External
Display. It hurts the policy which External Display use only
one MPP resource using GPU composition.
It changed pre-assigned Secondary Display information to Primary
Display when Secodary Display is not valid.

Bug: 238373481
Test: manual DP unit test
Signed-off-by: YongWook Shin <yongwook.shin@samsung.com>
Signed-off-by: Petri Gynther <pgynther@google.com>
Change-Id: Ia31ee867a92d133eaf831a0b2254433118fb8932