libhwc2.1: support blocking zone for min idle refresh rate
This change adds the support of voting min idle refresh rate
by different clients. The brightness can check whether it's
in the blocking zone and vote the value. Only the max value
will be set to sysfs node. The related properties should be
added in the device-*.mk file to support it.

Bug: 256662691
Test: change DBV and check logs
      enable/disable the proximity and check logs
Change-Id: I6f7bf1cec73331c7b6e0e34c6d7999c361119fcb