libhwc2.1 : reset assigned resource when a layer is destroyed
HWC service is crashed due to accessing memory which is already freed.
In destroyLayer, it only delete layer object and pointer in ExynosDisplay.
But ExynosMPP still have pointer of deleted layer object in mAssignedSource.
So when validateDisplay called, it refers the freed memory in resetAssignedState()

layer->resetAssignedResource() is added before delete layer object
so that remove layer info in ExynosMPP.

Bug: 264510390
Test: stop surfaceflinger by command "setprop ctl.stop surfaceflinger"
Change-Id: I5191ae334c7566c8fe503c06784aae0e6be1ea81