libhwc2.1: correct refresh rate reference at toggling lhbm
The refresh rate associated with mActiveConfig takes one or more frames
to actually becomre effective on hardware, and can't be used as reference
to decide if it's needed to set new refresh rate while toggling lhbm.

Use drm value instead, which can't be blocked by enableConfigSetting()
and takes 1 frame at most to reflect on hardware.

Bug: 280727927
Test: check udfps enrollment failure rate
Signed-off-by: Shiyong Li <shiyongli@google.com>
Change-Id: Ie2f39f73761cc6f9fe51637aa3ebf8e5d85026e5