libhwc2: HWC should check power mode when canSkipValidate()
In ExynosDevice::canSkipValidate(), HWC traverses all plugg-in
displays to check whether the validate() can be skipped or not.
However, the plugg-in but OFF display causes that canSkipValidate()
always return false because its mRenderingState is always
RENDERING_STATE_NONE.

For example, when un-folded mode, the OFF outer display causes
that the inner display's valiation() cannot skipped. HWC should
check not only plug state but also power mode.

Bug: 289167905
Test: swipe left/right at home screen, and check its trace
Change-Id: I8544c28c67fd533c1eefb8f4068883b1b0c9f684