libhwc2.1: revert behavior of setLayerCompositionType()
This is partial revert of ag/23598344. This will fix power regression
& disappearing cursor issues caused by that patch.

In ag/23598344, we didn't set mGeometryChanged if the composition
is changed back to the original composition (initially set by
setLayerCompositionType()) after falling back to client composition in
the previous frame. But, it causes some regressions, e.g. power jump
and disappearing cursor issues.

Thus, this will revert to the previous behavior i.e. to always set
mGeometryChanged as dirty whenever we change mCompositionType.

Bug: 289010449
Test: v2/android-power-te/functional_media_test_suite/*
Test: Use mouse and the cursor shouldn't disappear unexpectedly
Change-Id: If73dc0353c9e3ede93ff02e675aa879bc54287e4