libhwc2.1: get atomic message when getting invalid commit error
When we get an invalid atomic error, we usually need DRM ATOMIC message
to check it. Unfortunately, we cannot enable it by default, otherwise
the logs will be too many. We can add a logic to enable DRM ATOMIC
message and re-send the atomic commit with TEST_ONLY flag when we get an
invalid commit error. This will allow us to get the log for checking
without printing too many logs.

Bug: 301599195
Test: simulate invalid commit error to check log
Change-Id: I72d39d25c03e084eddf7b583291cfabf42a2793f