libhwc2.1: use frame scanout period to calculate BTS
Display frame scanout period is one major factor to calculate BTS.
Previously, we referred to it as the vsync period, but this is not
always true. minFrameIntervalNs was introduced to indicate the minimal
time that needs to pass before the next frame can be presented. We
should refer to it to calculate BTS if it exists.

Bug: 300082982
Test: refresh rate switch
Change-Id: I6340c401d0d6820c2d0d2a40f927cc93c5383008