libhwc2.1: Propagate content protection status to SurfaceFlinger
Kernel has a DRM property called "Content Protection" which is updated
whenever HDCP status changes. This CL adds this property to the
drmconnector in HWC and implements the logic to handle notifications of
DRM property changes being sent to user-space via kernel UEvents.
Also, since we cannot change HWC HAL interface at the moment, we'll use
a hack/workaround to notify SF about content protection change by
sending an onVsync callback with specially crafted parameters and gated
by a sysprop (similar to what we did for hotplug errors). In the future
we'll add a proper API in HWC HAL for this.

Test: manual
Bug: 307592398
Change-Id: Iea6b0c03aa180ea22ccc79f43421a97a9613d253