hwc3: Use the new HotplugEvent API
There's a new HotplugEvent API introduced in Android, that will allow
hardware composer to have a more straightforward path for reporting
hotplug errors, instead of using a hack/workaround in place today for
sending hotplug error codes to Surface Flinger via negative onVsync
callback. This CL adds the code to use the new API, if it's available,
but still keeping the old code to allow fallback to the old hotplug
API, if the new one is not available or not registered.

Test: manual
Bug: 310227347
Change-Id: Ib2047a3f2f604e11933df73366f326ac223da5c0