libhwc2.1: support set operation rate to min BTS requirement
When the DDIC is in HS mode, its operation rate will be high. If DPU
start to write data early, we can reduce the BTS requirement. If we do
not do this, we need to set a high BTS. Add support to set the
operation rate as the minimum BTS requirement if needed.

Bug: 305987242
Test: NS/HS switch, AoD in/out, in/off
Change-Id: Ib7837bc715bc4f638cd9aea5ab246bae0f2898b1