libhwc2.1: move all secure FB caches to a separate dedicated map
Secure buffers can still be cached in mCachedLayerBuffers. So the limit
number is still up to 32 buffers per layer, however secure memory limit
is much lower so we should move all secure buffers to a dedicated map
for secure buffers i.e. mCachedSecureLayerBuffers. So that we can ensure
to limit the secure layers up to only 3 buffers per layer.

Bug: 322343308
Test: face unlock enrollment, netflix, youtube
Change-Id: Idee9a34513c5880d96cd9bbcb9a5e51c93fc5774