libhwc2.1: Make ExynosDisplay::invalidate virtual
We need to clean up external display layer mappings whenever the
external display gets connected/disconnected. When ExynosDisplay
processes a hotplug event, the invalidate method is invoked. But
we need to make it virtual, in order to allow it to be overriden in
ExynosExternalDisplayModule. Also added a DisplaySceneInfo::clear
method for clearing layer mappings.

Bug: 325549729
Test: checked video colors on external display during HDR video playback
Change-Id: I1997d65a9954b85c4d811ddf95a38a45e49a7817