libhwc2.1: Implement dynamic decon switching
On foldable devices we have potentially 3 displays (inner + outer +
external), but only two of those can be active simultaneously. We also
have 3 decons, but only decon0 and decon1 support color management.
Previously displays were assigned decons at HWC startup and after that
decon assignment never changed, but that's problematic, because the
external display always gets assigned to decon2 and cannot support color
management. In order to fix that, this CL implements dynamic decon
reassignment in 3 cases:

1. When an external display is connected and we have a powered off
primary display (typical case).

2. When a primary display is being powered off and the external display
is still on decon2. This means the external display was connected during
active Dual Concurrent Display mode, at the moment of hotplug we
couldn't reassign decons, but now that a primary display is powering
down we can.

3. When a primary display is about to be powered on, but the external
display is currently using its decon. This means the phone was folded or
unfolded while being connected to the external display and we can try to
find a powered off primary display and reassign its decon to the
external display.

Bug: 329034082
Test: check color inversion on external display
Change-Id: I776abdf4e192ff2d894f811e4656d5d727479707