libvrr: fix residency accuracy for OFF and small render rate
There are some bugs causing inaccuracy for powerstats residency, i.e.
[1] Statistic present timeout event could be executed earlier than the
scheduled. Because this event could be dropped while
VariableRefreshRateController::threadBody() is still waiting to wake up
according to the old mWhenNs, meanwhile it should be updated to later
time.

[2] Statistic present timeout event should be scheduled according to
presentTimeNs instead of getNowNs() so that we don't have extra time
(around `presentTimeNs - getNowNs()`) recorded duplicately. It could
become innacuracy if the expected present time is quite later.

[3] When power mode is changed from OFF to ON, the added time to
mPowerOffDuraionNs could be wrong because it picked the wrong
mLastTimeStampNs.

[4] The first present after resume should be ignored since
mLastPresentTimeNs is recorded before OFF. So the first present should
set mLastPresentTimeNs to the correct value first.

Bug: 331440359
Test: hwc-tester, powerstats dumpsys
Change-Id: Ie24998c19a700f69393698adac064e2da101d3b3