hwc: add support for clearing buffer slots when display is off
For now, we would clear the buffer slots if the display is already off
such that it would not cause any flickering issue.

We would implement the proper solution after figuring out how to manage
the active buffer when the SF is still trying to clear the active
buffers.

Bug: 330806421
Test: switching display, open/close apps
Change-Id: I4b8aa64708fb6d78ba9e354e3a68c6d589f83e38
Merged-In: I4b8aa64708fb6d78ba9e354e3a68c6d589f83e38