libhwc2.1: first frame insertion timer
Previously, the first frame insertion timer was 33.33 ms; however, the
new frame insertion table has changed it to 16.67 ms. Furthermore, we
may need different first frame insertion timers in the future.
Here, we implement the API to read the settings and set the first frame
insertion timer accordingly

We also relaxed the confidence percentage of the video frame rate
estimator from 95% to 50%

Bug: 338527271
Test: Change the first frame insertion timer & verify its effect in trace
Change-Id: Id5958ac5a985f40711b2a2a6cde336b8a473f53f