libhwc2.1: libvrr: fix peak refresh rate issue for min refresh rate flow
We recently added the flow round indicator to validate the refresh rate,
which rounds the refresh rate to the closest valid value as ag/27391746.
This flow takes effect when the configuration changes. But it wrongly
uses the old configuration and trims the correct refresh rate (120) to
the maximum refresh rate of the new configuration(60).

This change fixes the timing so that the correct refresh rate will be
reported.

Bug: 344478250
Bug: 344478264
Test: test manually according to the steps in the bugs.
Change-Id: Ic2fecc3791397b996687ac7e884971cd78ab770e